refactor(test): migrate auth.ts helpers to three-tier test architecture
Move createTestCliToken, deleteTestCliToken, createTestDeviceCode
to db-test-seeders/auth.ts with @why-db-direct JSDoc.

Move findTestDeviceCode, findTestCliToken to db-test-assertions/auth.ts.

Re-exports in api-test-helpers/auth.ts preserve backward compatibility.
No test files modified.

Closes #9358
